+++ src/Readarr.Api.V1/Diagnostics/DiagnosticsPushResultResource.cs
@@ -1,0 +1,10 @@
+namespace Readarr.Api.V1.Diagnostics
+{
+    public class DiagnosticsPushResultResource
+    {
+        public bool Success { get; set; }
+        public string Message { get; set; }
+        public string Commit { get; set; }
+        public string Folder { get; set; }
+    }
+}